improved time to sleep when no bucket is available
The 'sleep_time' of a bucket say how long the worker should sleep before trying to get the next
bucket if there was no bucket available.

This time was 'until_next_iteration' but during dictatorship this could be a very long time. And
while the dictatorship ends as soon as the leader has computed/updated all buckets the workers are
still sleeping very long.

Now the sleep time is max the half of the iteration_length which is also the maximum time the leader
is sleeping before updating the buckets.