General cleanup and switch to Ruby 1.9
The MDNS functionality was dropped since net-mdns is not compatible to Ruby 1.9 and dnssd did not reliable work.
It was used only for find_workers due to graceful stop anyhow. This has now become a callback.
If graceful stop is needed a working implementation of find_workers has to be present.